requirements.txt
fastapi
uvicorn[standard]
aio_pika
asyncpg
databases
sqlalchemy
pydantic
aiosmtplib
python-dotenv